Develop the domain models for the TODO APP
We'll have zero or more lists made of zero or more items. The items may
have sub-items (for splitting complex tasks, for example) and they can
also have tags for categorization.

A tag might be a meta-tag for more flexibility, brevity, and
expressiveness. For example, the usage of a `java` tag could have the
same meaning and categorization properties of both, the `development`
and `jvm` tags.